RepositoryFinder to work with linked folders
When displaying the path to a repository contained within a linked
folder, use the path of the linked folder within the project as
the prefix of the repository location as the repository location
string is relative to that folder.

When a tree node is selected without a project, use the parent tree
node to acquire the project.

Change-Id: If7214ca8bf10439e7e27b424864742a10dd8b766
Signed-off-by: Shawn O. Pearce <spearce@spearce.org>