Update patch set 1
Patch Set 1: No score

If we're going to get anal about tabs-or-spaces in the plugin.xml, we should set the policy to force one or the other. The items looked (on my machine) formatted correctly, but that's because my workstation is configured to put tabs in when I hit the tab key. If we don't want that, we have to explicitly set this in the project's .settings directories so that everyone's coding from the same stylesheet.

Patch-set: 1