Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

My problem with this change is it makes it very hard to selectively edit part of a file which already has trailing whitespace on its lines.

E.g. many of our source files have trailing whitespace in Javadoc comments ("\t* \n" are empty comment line).  If you only modify a couple lines of code in a method, and save, this triggers Eclipse to rewrite most of the file fixing those comments.

Are we going to clean up the existing source files first?

Patch-set: 1