Update patch set 1
Patch Set 1: No score

It's possible to choose between the whole file and modified lines for formatter. But not for additional actions like removing trailing white spaces.

It would be better to clean up the existing source files now when there are not so many contributions, these changes are trivial and they wouldn't affect many people.

Uploading patches and reviewing them several times just because of the trivial style issues seems to me not a best workflow ;-)

So my POV on this issue:
- Enable save actions 'Format edited lines', 'Remove unnecessary blocks' and 'Remove trailing white spaces on all lines'.
- Apply these rules to the existing source files now.

Patch-set: 1