Update patch set 3
Patch Set 3: No score

> Test scenario 1.)
>     * Team > Share Project > Git -> Next -> Select test2 and click "Create"
>     * > this initializes test2/.git but project is not shared
>     * again Team > Share Project > Git -> Next
>     * > test2 now listed to have repository C:.git (invalid path!)

As I can see original behavior has been preserved, as to be compared with commit 07081c040d. It seems to me this inconsistency deserves to be filed as a separate bug, because this patch doesn't address this issue, has touched several places and is already big enough.

>     * select test2 and click finish
>     * > project is shared, repo is at test2/.git
>     * adding class Hello.java and committing the change works

Patch-set: 3