Update patch set 4
Patch Set 4: Looks good to me, approved; IP review completed

I agree with Mykola, we should fix this now with this patch, and work with the other issues in a different bug, and a different change.  Basically this whole share wizard stuff has a horrible UI and is broken in a bunch of ways.  But this change is at least a step in the right direction by trying to keep track of the project<->repository mapping data.

Patch-set: 4