Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

This depends on the abandoned version of the same change.  You  need to squash the commits together so the entire thing is only one commit.

Also, our convention is "Bug: 291133\nSigned-off-by: ..." in the footer.  We don't have it yet but I want to have our gitweb and Gerrit UIs link back to Eclipse Buzilla based on that pattern.

Patch-set: 1