Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

In my opinion the "Compare with Revision" feature should open a dialog. If just the History View opens the user does not directly know what to do further.

In the History view I would prefer a menu entry "Compare" which compares with the working tree in case one commit is selected and which compares two revisions in case two commit are selected. But I could also live with the double-click solution.

All in all I think that you should split the patch into three patches which could be discussed independently:
- the fix for the GitCompareFileRevisionEditorInput
- the new CompareWithRevisionAction
- the new behaviour of the GitHistoryPage

Patch-set: 1