Implement GIT_CEILING_DIRECTORIES for eclipse.
If this environment variable is set it will be interpreted as a list
of directories above which which the Git Team Provider does not search
for repositories.

The key here is "search". By explicitly pointing at a repository it can
still be used, regardless of where it is located.

This feature is mostly useful for testing, but can possibly be used to cut
of scans into network directories, which is the original rationale for the
implementation in C Git.

Change-Id: I0d39eaa63670030b495148a4cf2a57add7185f89
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>