Update patch set 1
Arrrgh, that revwalk patch comes back to bite me almost 2 years later.  Yikes.

(For those who aren't sure, the two commits Robin is citing are from our old history, before we moved to Eclipse.org, and must be fetched off repo.or.cz if you are interested in reading them.  But don't because 92a1cc82 is a giant 6811 line deletion, which happened to remove the feature this change is trying to put back.)

Patch-set: 1