Update patch set 2
Ah, some light...

As I see it this action should open the history view so the user can then compare a selected revision with the local version, which we cannot yet do. This this action promises
too much. This original view had this, though it only worked well on Linux for some mysterious reason.

It should also be named just "Git Revision..." (or possibly "Revision..." to avoid the strange looking 

"Compare with... Compare with Revision" menu.

Patch-set: 2