Update patch set 2
Patch Set 2: No score

This behaves like "Compare with->History..." in the CVS module. So I think the user would expect to see a view with the revisions of the file. I do not know GIT terminology that well because of my CVS background. Maybe we should rename the menu entry to make it clearer.

Patch-set: 2