Update patch set 2
Good point, I admit that I'm not that familiar with CVS. I also see the advantage with this solution that we do not need to duplicate the list of revisions in a new dialog.

However, I suggest that we also use the same label then as CVS: Compare With -> History.

Patch-set: 2