Update patch set 2
I also first thought we should bring up a dialog when the user selects "Compare with Revision".

But together with http://egit.eclipse.org/r/122 (Compare with the working tree on double click) the behaviour would be the same as in CVS. We should however also call the action then "Compare with History..." as CVS does. Then the user would not be astonished that no dialog comes up.

We should also add "Compare" actions in the History View. I already did that in my repo but wanted to wait until this change and http://egit.eclipse.org/r/122 are thoroughly discussed.

Patch-set: 2