Update patch set 8
Patch Set 8: I would prefer that you didn't submit this; No score

Matthias, can you update this to depending on the jgit bundle instead of the feature in the feature.xml?

<import plugin="org.eclipse.jgit" version="0.6.0" match="greaterOrEquals"/>

The JGit bundle shouldn't be produced as part of the update site build, that is handled by the JGit build.

Patch-set: 8