Resolve JGit dependencies via JGit P2 update site
Option 2 to remove the mix of P2 and Maven dependencies
between EGit and JGit build.

The build of the update site assumes that EGit and JGit are checked out
under a common root folder, this is needed to refer to the JGit update
site. If the layout is different the URL pointing to the JGit update site
has to be passed to the maven build as system property jgit-site, e.g.
   mvn -Djgit-site=file:/C:/tmp/site clean install
or
   mvn -Djgit-site=http://point.to.jgit/site clean install

This change depends on the JGit change I7f3638ccc512be16ed13ab1d726773d2cf779f2a

Change-Id: Ib9fb6c238582e7d2a0f63f39596a6e6af91546bb
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>