Create patch set 3
Patch Set 3: No score; No score

This one is still based on the old(?) swtbot, but that is the only one stable version I've had.

We need to make it work with org.eclipse SWTBot, including either JUnit 3 or 4 in such a way that it is possible to re-Run single tests.

Patch-set: 3
Subject: Create an automatic UI test for the Git Import Wizard
Commit: be23d88a1f99419b913bfd991af269474eccdd01