Changed the project import page checkboxtable into a filtered tree.
With large repositories it is difficult to get a subset of projects because the
import projects tree could not be filtered. Now the import projects uses a
filtered tree which is commonly used inside Eclipse.

- The tree can be filtered with *text like syntax.
- If a project is selected it always stays in the view even if it does not match
  the filter.
- If projects are deselected with the "deselect all" button they can be removed
  from the view if it does not match the filter string. This is done by calling
  setinput again. This is not done for individual deselections because that did
  not look right and I am not sure about the setinput overhead on really large
  repositories

I have tested this with the following repositories:
git://github.com/smilebase/org.eclipse.mylyn.github.git
git://github.com/aptana/studio.git
git://ecf1.osuosl.org/home/gitcvs
and the egit and jgit repositories

It was noticed that some of the aptana and ecf projects (notably the feature
projects) were not attached to teaming after import. This is probably not caused
by this change.

Change-Id: I6fb40104863323aaf66ea355f3574e08fa775b17
Signed-off-by: Wim Jongman <wim.jongman@gmail.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>