Update patch set 4
Patch Set 4:

Hi Robin,

I have to admit that I have not yet tested the code on linux (it is a flaw that we do not yet have linux and Mac OS instances in our department).

However, the code is basically the same as in CommitFileDiffViewer (l.62). 

The double click only has effect when the input of the history page is file. Could you please check once again if in your case the history is shown for a file and not for a folder or the project?

Thanks,
Stefan

Patch-set: 4