Add compare actions to history view.
Add two actions to the context menu of the list of commits
when a file is selected. Comparing folders is not
implemented yet.

If one commit is selected, compare the content of the
file with the working tree, if two entries are selected,
compare the two contents of the file with each other.

Bug: 301948
Change-Id: Ibb2382b19fead8e3cec215823db60028532d076b
Signed-off-by: Stefan Lay <stefan.lay@sap.com>