Parse .gitignore to find file's ignore status
Used class IgnoredResources.

Method from this class is called in IgnoredAction, replacing the check
to Team.isIgnoredHint.

Currently:
- Calls Team.isIgnoredHint
- Checks .gitignore, using the special rules and cases for git

From what I can see, there are two places in the code that check
Team.isIgnoredHint, IgnoredAction and DecoratableResourceAdapter. The
latter has a TODO that says it should check for .gitignored and
.git/info/exclude. I've only added a call to the new isIgnored
to the IgnoredAction for now.

Also I really hope this git push works :( I wasn't sure how to amend
the previous patch, and my tree was a bit dirty so I did a fresh
clean and copied the files over. Sorry if it causes another mess!

Change-Id: I5c40e237275f2d9031b2b56591ba98c5ce166008