!rse .gitignore to find file's ignore status
Used class IgnoredResources.

Method from this class is called in IgnoredAction, replacing the check
to Team.isIgnoredHint.

Currently:
- Calls Team.isIgnoredHint
- Checks .gitignore, using the special rules and cases for git

From what I can see, there are two places in the code that check
Team.isIgnoredHint, IgnoredAction and DecoratableResourceAdapter. The
latter has a TODO that says it should check for .gitignored and
.git/info/exclude. I've only added a call to the new isIgnored
to the IgnoredAction for now, so compatibility can be tested by
right click -> Add to .gitignore and checking whether or not an
addition has been made to the relevant .gitignore file.

Currently should be compliant with .gitignore, a list of
compatibility tests I've done will be posted in the comments.

Change-Id: I5c40e237275f2d9031b2b56591ba98c5ce166008