Update patch set 2
Patch Set 2: Looks good to me, but someone else must approve; IP review completed

I made the changes requested.

We ship with the wikitext jars because this is the pattern other projects use. I would like to get this in before we release 0.7.0... we can always make changes in the future.

Patch-set: 2