Update patch set 1
Patch Set 1: Abandoned

This seems to have been created at my first attempt to push http://egit.eclipse.org/r/#change,325. I had not seen an effect on that change so I rebased the commit and pushed again. Now everything seems to be ok, and this change is not necessary.

Patch-set: 1