Update patch set 1
Patch Set 1: No score; No score

The problem with this approach is that we want to manually generate the doc instead of doing it on each build. The reason for this is that it's possible to easily break the build with doc modifications. The Mylyn project switched to this approach. However, I guess we can be OK with it to invoke it everytime.

Patch-set: 1