Allow to use the repository name in Git Label Decorations
Currently, the Git (text-) label decorations include variables for dirty,
staged, and branch. Since there are many repositories with the same "master"
branch, we should have a "repository name" variable. This should be the name
of the repository's directory. While not necessarily unique, this helps in
identifying the repository in the project explorer.

Bug: 306397

Change-Id: I72be529419748c43114538edae8fafd811c38f55
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>