Show a "Create Patch" context menu entry in history view
When one commit is selected in the commit list, the diff to
its parent can be saved to a file, together with meta-information
needed to send the patch by mail.
    
The file filter of the history view is used. The diff written to
the patch file is the same as that shown in the history view.
    
It is not possible to create patch files for the first commit and
for merge commits.

Bug: 297636    
Signed-off-by: Stefan Lay <stefan.lay@sap.com>