Use JGit from the nightly update site
It's not convenient to have JGit projects be opened when developing
EGit-only stuff. Extend target platform definition with a latest JGit
from the nightly update site [1] listed on 'EGit - Download' page [2].

Nightly update site used despite of the fact that it is intended to be
updated frequently and JGit's version here should be kept in sync.

[1] http://download.eclipse.org/egit/updates-nightly
[2] http://www.eclipse.org/egit/download/

Change-Id: I6e8fce987d9c107078c2f4e0d798dd724099f451
Signed-off-by: Mykola Nikishov <mn@mn.com.ua>