Update patch set 1
Patch Set 1: No score; No score

I would propose that we sort the list in that way that the latest entry will occur at the first position, even if it had been in the list before. When there are more than 10 entries we should throw out that one that was used the longest time ago.

Patch-set: 1