Minor Cleanup
This is about cleaning up the anti-pattern
of constructing an Exception in a utility
method and throwing it in the code. This
will result in wrong stack traces (found
both in EGit Core and EGit UI).

Furthermore, a couple of methods were
placed in the classes that need them instead
of having them in the plug-in activator (only in
EGit UI).

A few places were marked as TODO where
an error status is created but neither logged
nor thrown as CoreException.

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>