Make tracing plug-in local
Sharing trace classes between plug-ins
is bad, as the OSGI tracing infrastructure
will write the ID of plug-in that exposes
the trace class into the trace file.
Instead the ID of the plug-in that is
writing the trace should be written.
This is achieved by obtaining a separate
Location for each plug-in.

Bug: 306793

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>