Show "Create (Git) Patch" context menu entries in history view
The actions are shown in history view and not in the package
explorer as in CVS. The reason is that when yiou use a distributed
versioning system you normally commit your local changes
to a local branch before you create a patch.

When one commit is selected in the commit list, the diff to
its parent can be saved to a file, for the "Create Git Patch"
action together with meta-information needed to send the 
patch by mail.

The file filter of the history view is used. 
    
It is not possible to create patch files for the first commit and
for merge commits.

The patch format is different for CVS and for Git. Therefore
there are now two actions. In the next iteration there will be
a wizard as in CVS where the destination and the format can
be chosen.

Bug: 297636    
Signed-off-by: Stefan Lay <stefan.lay@sap.com>