Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I think the terminology needs to be changed to fit in with other Eclipse team providers. If I wanted to undo such a change with SVN or CVS, I'd do Team -> Replace With HEAD (or Replace With Version). Wouldn't that have the same effect here?

There are some messages which we need to internationalise in the error dialogs as well.

Patch-set: 1