Update patch set 4
Patch Set 4: I would prefer that you didn't submit this; No score

I tried again the same test, "Replace with HEAD" now replaces the selected file with the HEAD revision but the label decorators in the package explorer still mark the file replaced by HEAD as dirty (">" in front of the file name). When I try to commit I get the correct question that there is no change and if I want to amend the previous commit. So something is wrong with the label decorations after running DiscardChangesAction.

Patch-set: 4