Update patch set 6
Patch Set 6: I would prefer that you didn't submit this; No score

Very useful patch.

However, it does not replace with the HEAD but with the git index. I would suggest to put the action completely out of the Team menu but to the "Replace With" sub menu, in analogy to the "Compare With" -> "Git Index".

When the project is in a conflict state there is an exception. Maybe you could check this before?

Patch-set: 6