Removed checkedItems field from GitProjectsImportPage; added a getCheckedProjects method instead, which computes the set of checked projects based on information available from the UI widgets.
This fixes at least one bug that was affecting my work: unchecking projects by using the spacebar didn't work, because there was no KeyListener in place to update the checkedItems field. This lead to the project being unchecked in the UI, but still considered checked by the wizard page, and upon clicking Finish, the wizard would try to import all of those unchecked projects.