Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

Also I'd prefer that you format the message with slightly longer lines. Please tell if the dialoge suggests this line length in your Eclipse. In my Eclipse, breaking the lines within the dialog it would look like below:

Make tracing plug-in local

Sharing trace classes between plug-ins is bad, as the OSGI tracing 
infrastructure will write the ID of plug-in that expose the trace class 
into the trace file. Instead the ID of the plug-in that is writing the trace
should be written. This is achieved by obtaining a separate
Location for each plug-in.

Patch-set: 1