Update patch set 3
Patch Set 3:

I mean that, no reaction on double click on tree element in Linux is an SWT issue and it should be fixed in SWT not in EGit.

Expanding tree elements on double click is default action not only for Windows but also for other systems. I don't want to remove it, I only want same behavior on other systems.

Patch-set: 3