Update patch set 5
Patch Set 5: No score

I've never used EGit under Windows so it is possible that those features already works on it. I'll check that and if it is already working I'll abandon this change.

From the other hand it is strange that SWT has so many differences (or rather lacks of usability features) between Windows and Linux.

Patch-set: 5