Add tagging UI support
Implements UI for unsigned and annotated tags. This implementation
allows the user to select which commit object should be associated with
created tag. List of all commits is available in the 'Advanced' section
of the dialog. There is a special combo widget that can suggest commits
based on match of typed-in text with commit SHA-1 or part of first line
of existing commit messages.

User can also edit/overwrite tag message or associated commit of an existing
tag by selecting it from the list of existing tags and marking the
'Force replace existing tag' check box.

By default tag is created for current HEAD, this can be changed in the
'Advanced' section.

The tag dialog is available in project's context menu ('Team->Tag...').

CQ: 4073
Bug: 311262

Change-Id: Icb30655845d1e1198e59992a10148421ab9f9bd5
Signed-off-by: Dariusz Luksza <dariusz@luksza.org>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>