Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

I usually avoid EGit UI stuff.  The code seems sane from a JGit API usage perspective, it should create a tag and store it.  My -1 score here is because there are some obvious things to be cleaned up before we can submit this, but its at least heading in the right direction and shouldn't be hard to polish up and submit.

Patch-set: 1