Show untracked files in commit dialog
Adds support for untracked files to the CommitDialog. Per 
default the files will not be shown in the list. Adds a new
checkbox "Show untracked files". 

Only files are shown which are not ignored by
Team.isIgnoredHint. The .gitignore is still not taken into 
account because there is not yet an API for it (see pending 
patch http://egit.eclipse.org/r/#change,326).

Bug: 303373
Signed-off-by: Stefan Lay <stefan.lay@sap.com>