Update patch set 1
Patch Set 1:

The easiest way is to create a new branch on top of egit's master branch. Then cherry-pick your changes into
it and finally use rebase -i to squash them together and update the comment so it looks like there ever was one patch. I suggest you re-use the change-id of the first
patch and abandon this one. Something like

Patch-set: 1