Fix incorrect editor updates with changes from diff-viewer
The issue occurs if BlobStorage.getFullPath returns a path
that is equal to the path of the file in the workspace. In this
case Eclipse seems to share the editor documents.
The fix provides new implementations for getFullPath
which provide a path of form 
<repositoy name>/<path> <commit id>
When viewing a revision this path is shown as tool tip
of the editor title (similar to CVS).

Bug: 302145
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>