Update patch set 2
Patch Set 2:

Please make sure the Change-Id is in there in the future. EGit will have a feature for this soon, but for the time being  you should install the hook (see wiki) for C git. Amending an existing commit from C git will add the change-id if the hook is installed. If you already pushed without a change-id, insert the changeid from Gerrit manually and push again.

Patch-set: 2
Label: SUBM=+1