Implement compare mode toggle in history view
The compare mode toggle defines the action to be executed on double
click. With compare mode enabled the compare editor is launched.
With compare mode disabled an editor is opened to view the selected
revision. Executing Compare->With history... launches the history 
view with enabled compare mode (similar to CVS).

Bug: 306779
Change-Id: Ie276191847b1e070dcd361606d6fc3f36468b318
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>