Remote configuration wizards
There should be some UI to allow specifying
remotes instead of having to edit the config
file directly. Some context menus were added
to the Repositories view to allow just that.
In order to re-use the existing wizard pages as
much as possible, there was some refactoring
required for these, too.

Bug: 304182

Change-Id: Ie17241dcf9a3665f899c6a044fc84afd04ceadb1
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>