RepositoriesView: use the GitImportProject wizard
The generic ImportExistingProject wizard from the 
org.eclipse.ui.ide plug-in does not allow to specify
a directory in 3.4.
Instead of the generic ImportExistingProject wizard,
the RepositoriesView should use the Git wizard which
allows to limit the scope to a directory in the file system
and also adds project filtering capabilities as well as
the option to add the imported projects to Team support.

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>