Update patch set 2
Patch Set 2: I would prefer that you didn't submit this

I suspect that the perspective is not available.

You could try printing out what perspectives are available via bot.perspectives() and get an idea of why that perspective is unavailable.

Could be a language setting or a missing plugin that contributes the perspective.

I'll work on improving the error message so it throws a WidgetNotFoundException instead of an Arrayexception.

Patch-set: 2