Update patch set 1
Patch Set 1: No score

I would rather have the branch at the beginning and the repo afterwards - the branch is what interest me more most of the time.

I also do not like the - as separator, it is not intuitively clear that it separates two different entities. How about braces for the repo: [branch (repo)]?

Patch-set: 1
Label: SUBM=+1