Git Repositories View: Import Projects
Add some context menu in the view to allow starting 
of different wizards for project import (not only the 
Import Existing Projects wizard).
This is also in preparation for a better Git Import wizard.
The bug below is not solved by this, but this should
help in the discussion of it.

Bug: 281394

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>