Update patch set 1
Patch Set 1: No score; No score

>for HEAD you give the full ref:
>refs/remotes/origin/master
>for other remote branches you leave out the refs/remotes
>e.g. origin/stable-0.7
>this looks inconsistent to me

One of the first requirements was to show the "shortened" name for branches, e.g. instead of refs/heads/master just master in order to imprvoe readability of the tree. This confused people because there could be refs/heasds/abc and refs/remotes/abc (and perhaps also a refs/tags/abc) which would result in three nodes with the same text. Thus separation by the remote/local/tag nodes. However, in case of showing the reference HEAD points to, we need to use the full name to avoid ambiguities.

>maybe the remote HEAD could be also marked with the little
>check mark like for local branches ?

What is the remote HEAD?

Patch-set: 1