Branching in RepositoriesView
This is a very first implementation of adding/deleting
branches in the RepositoriesView. It works on the
branch-related nodes.
This has no extensive input validation or such, and
it also does not take other repository states into
account, so perhaps this might have to be
added, too.
The "Delete Branch" operation uses "force"-mode
to make sure it is not rejected if the branch is not
a parent of the current branch or such.

Bug: 311354

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>