Update patch set 2
Patch Set 2: No score

Hi Remy,

did you see my comment on your comment? It was:

We do not know which projects will be affected by the branch operation. It's e.g. possible that changes in a project are performed that resides in an EGit repo but the project is not shared with EGit. Thus I think it's safer to lock the complete WS.

Best regards

Jens

Patch-set: 2