Git Import Wizard
This is the new Import wizard for importing Projects from Git.
Similar to CVS, it first shows a list of Git Repositories. The user
may select one from the list (the same list as shown in the 
Git Repositories View) and continues from there. Alternatively,
the user can clone or add a Repository to the list.
Then all continues like with the Import Projects... wizard
used in the Git Repositories View.

Bug: 281394

Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>