Update patch set 6
Patch Set 6: Looks good to me, approved; IP review completed

This is great work with Mathias.

I found one issue with the UI and it's that the wizard opens with an error message. This is against the Eclipse UI guidelines and we should avoid this. I opened bug 311834 for us to fix this issue. Otherwise, we should push this now :)

Patch-set: 6
Label: SUBM=+1