Update patch set 3
Patch Set 3: No score; No score

I like the direction this is going.

I recommend that on the first wizard page, "Clone a repository" and "Add local repository" become buttons and are moved to the right side of the list of git repositories. At the moment, they are simply located below the list of repositories which isn't optimal imho.

Patch-set: 3