Update patch set 5
Patch Set 5: No score; No score

If I try to import the parent folder of an already imported project as a general project (which of course isn't possible) I get an error in the error log and a nasty error popup. Instead I would prefer to be notified in advance that this will not work.

Patch-set: 5