Update patch set 1
Patch Set 1: No score

Would it make sense to also make this work if the immediate parent folder of the repo is pasted ?

Would drag&drop from the filesystem also be possible ?

The change needs to be rebased. Maybe you should switch to do patch series instead of several overlapping changes, this may reduce the need for frequent rebase and conflict resolution cycles since patch series are always released according to the order of the series. If the changes are posted independently the submit order depends on which changes pass review faster.

Patch-set: 1