Prepare EGit for the new State MERGING_RESOLVED
JGit will report a new repository state MERGING_RESOLVED. Currently EGit
doesn't support committing in this state. Missing is the correct handling
of the content of .git/MERGE_MESSAGE and .git/MERGE_HEAD. When committing
in this state these contents have to be used and after a successful commit
these files have to be deleted.

This fix adds a special check for this new state and disallows committing
in this state.

This Change depends on JGit Change http://egit.eclipse.org/r/#change,667.
This change will break the build of EGit if this JGit change is not merged.
On the other hand: As soon as the JGit change has been merged we'll have to
merge this change soon.

Signed-off-by: Christian Halstrick <christian.halstrick@sap.com>