Update patch set 2
Patch Set 2: No score

Provided rule in ResetAction.
Remy, what do you think about the rule for TrackOperation?
TrackOperation does not write resources. It reads resources and adds them to the Git index. If we do not lock the files an other thread might e.g. delete the files and the TrackOperation would fail. So should we lock the files when performing TrackOperation?

Patch-set: 2