Branch Selection Dialog UI improvements
The dialog shown for Team->Branch... and Team->Reset To... was
reworked to re-use the same tree as the Git Repositories view.
Also, double-click is now supported to perform the
"Ok"-Action.
Furthermore, we found a problem with checking out remote
branches: HEAD is pointed to the remote branch instead
of to the underlying commit; for the time being, we deactivate
check-out of remote branches

Change-Id: I68eb12f7c6d7b46488650e635639772a95a88e90
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>