Show "Create Patch..." context menu entry in history view
The menu entry is shown in history view and not in the package
explorer as in CVS. The reason is that when you use a distributed
versioning system you normally commit your local changes
to a local branch before you create a patch.

When one commit is selected in the commit list, the diff to
its parent can be saved to a file or stored in the clipboard.

The file filter of the history view is used.

It is not possible to create patch files for the first commit and
for merge commits.

Per default the wizard creates a patch with a format which can
be applied with the eclipse apply patch wizard:  The paths to the
files have no prefixes (as with git format-patch --no-prefix) and
they are relative to the eclipse project and not to the repository.

There is an option to create a patch which can be applied with
"git apply" on the command line (Not yet in eclipse.).
There may be more options in the future on the Options page.
Before more options are offered the diff rendering code should
be moved to jgit.

[ms]
- fixed some style nits
- reduced initial wizard height

Bug: 297636
Change-Id: I8d53a4c7685df75887ad6ec80aeda22dbb31e01f
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>