Update patch set 5
Patch Set 5: No score; No score

I got the feeling that this is spinning out of control, so I reverted to master and re-applied only the changes needed to replace the tree and fix the "New Branch" issue.
Fortunately enough, in conjunction with change 680, check-out of remote branches is working now.

Patch-set: 5