Update patch set 1
Patch Set 1:

Hmm... You simply amend your commit and make sure to have the Change-Id: <this change's ID> in the commit message (here, the ID is I5c0889e9f6e95c2c62d96195623499802ce701a1, see header). Not quite sure what are allowed places for this line, I always put it before the Signed-off line.
Not sure why you would need a force. If the ID you provide in the commit message is found in Gerrit, a second Patch set is added to this change when you use the push command as described in the contributor guide. Did I miss something?

Patch-set: 1