Update patch set 6
Patch Set 6: No score; No score

Hey guys, not to be the evil pragmatic guy here, but can we get a minimum level of functionality that's acceptable and then get the CQ moving on this feature? Ideally I'd like to get this in for 0.8.0 and would prefer the CQ started soon. We can always polish this feature in the future without the need of a CQ :)

Patch-set: 6