Update patch set 3
Patch Set 3: No score

It seems to me that TagInputDialog, which has code copy/pasted from BranchSelectionDialog, suffers from the same problem that's described change 580.
http://egit.eclipse.org/r/580

Also, I think it would make more sense if there were two separate actions for tagging and branching instead of stuffing them both in the same dialog. If they both need to go in the same dialog, then I think we need to put the branch/tag/rename buttons off to the right of the table instead of having all of them at the bottom. Five buttons is a lot for the bottom of a dialog and I don't think I've seen that anywhere in the Eclipse SDK.

Patch-set: 3