Update patch set 8
Patch Set 8: No score; No score

Chris: In a hurry? Don't panic! A problem with problems in UIs is that a small one can take a lot of code to fix and if we push it through a CQ, we'll need another CQ to fix it, and a dozen bug reports to tend to. It needs to be good enough for anyone to fix without resorting to duct tape.

It's better with a nice working one commit Dariusz can be pround of and at least doesn't *have to* be fixed. 

We might think of better UI's later on but that's another story.

Patch-set: 8