Update patch set 7
Patch Set 7:

If you select a tag to see its message the I cannot create a new tag anymore. I can select "advanced" and "force replace existing tag", and do it that what, but that is somewhat strange. 
(this is not a showstopper, but we will get bug reports, I'm pretty sure)

Patch-set: 7