Update patch set 6
Patch Set 6:

Dariusz wrote: In my humble opinion writing UI unit tests before we establish how UI should look like, behavior and what features should be implemented is worthless. Because that I haven't implement them as well as documentation for this dialog but I have both of them in my mind. I'll move on with documentation and junits when this dialog will be concrete.

That's why I used the term "consider" :)

Patch-set: 6