Update patch set 8
Patch Set 8: No score; No score

great we are getting close

tried on Windows XP:
- now I am able to replace existing tags and OK button behaves properly
- clicking "Advanced" only shows up the label "Choose commit that should be associated ..." 
but no combo, resizing the dialog reveals that this is a layouting problem. 
Looks like the control on which the advanced section is expanded doesn't resize sufficiently 
to show the combo for selection of commit to be tagged.
- clear button got no text for its label in the resource bundle
- the initial dialog width looked a bit broad but that could be caused by a long error message being displayed instead of the missing clear button label

Patch-set: 8