Update patch set 1
Patch Set 1: Looks good to me, but someone else must approve; No score

Yes, that is annoying. I guess we couldn't make our minds up. For the time being this change is ok. 

This variant works well with the fact that clients construct collection and a collections make a smooth interface, while arrays seems to be more "eclipse"-style.

Patch-set: 1