Update patch set 2
Patch Set 2:

I think if an untracked file is selected, then that option should be checked and then it can just be preselected like any other file. Unchecking and checking it back on won't select it again but I think the behaviour of ticking the checkbox if something is untracked seems like something the user would want.

Patch-set: 2