[RFC] Add a 'Staging' view to streamline the workflow of staging changes
Git's staging area is one of its most powerful features. This view
provides users with the ability to drag-and-drop files between
nodes to place modified or untracked files into the staging area.
Such an operation would be synonymous to adding the modified files
to the index and tracking the untracked files. Other operations
like committing is also possible from the view's context menu.

[ms]:
- rebased to current master
- fix unhandled InvocationTargetException in StagingView.java

[bl]
- rebased original implementation from Remy (I5fb81a1811cf6f2aff22016438c664fc5014b9a5)
- initial attempt to remove deprecated api usage
- requires jgit change I43c13eb72a44f80135c93525fce0c0280b0e64a2 for
  conflict support in IndexDiff

Change-Id: I7d78aff4db32d864ac0410a178bad149c45b6440
CQ: 4125
Bug: 313263
Also-by: Remy Suen <remysuen@ca.ibm.com>
Also-by: Matthias Sohn <matthias.sohn@sap.com>