Add a 'Staging' view to streamline the workflow of staging changes
Git's staging area is one of its most powerful features. This view
provides users with the ability to drag-and-drop files between
nodes to place modified or untracked files into the staging area.
Such an operation would be synonymous to adding the modified files
to the index and tracking the untracked files. Other operations
like committing is also possible from the view's context menu.

[ms]:
- rebased to current master
- fix unhandled InvocationTargetException in StagingView.java

CQ: 4125
Bug: 313263
Change-Id: I5fb81a1811cf6f2aff22016438c664fc5014b9a5
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>