Update patch set 2
Patch Set 2: No score

I misinterpreted the code. I think we need to warn the user here. If we cannot read the index, that is a pretty fatal operation and they need to know about it. Giving them a compare editor with blank content is not a good idea as it may give them the impression that their index is corrupted and that their staged changes have been lost.

Patch-set: 2