Update patch set 1
Patch Set 1: I would prefer that you didn't submit this; No score

Concerning the IOException: Would you think that in this case we should not bother the user with a dialog? Then we should maybe also change the CompareWithHeadAction. But I think that we should at least log it to the error log.

> One problem is that the editor's title is just "Compare X Current and Current" when doing this. 

I haven't even noticed that, we should however fix it with another patch.

Patch-set: 1