Git Label Decorations of "detached" HEAD
In the case of "detached" HEAD, currently, there is only
the commit ID (truncated) shown in the Project label decoration.
This fix implements the suggestions collected in a mail thread
on the egit-dev mailing list by adding the information which
tag or remote branch is pointing to the currently checked out
branch.

Change-Id: 	I70c791f9552c6201b1c66488ab45f8186ec4c28d
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>