Update patch set 1
Patch Set 1: No score

One thing to note is that this code will only work if a checkout was performed using the BranchAction. This will not fly if the user were to checkout another branch from the command line.

Though technically speaking, that is a general problem (see bug 313046).

Patch-set: 1