Create change
Patch Set 1: No score

I realize that I've left the code in the UI bundle but I wanted to get people thinking about how to refactor the code so that it can be reused more easily. Having to create the action directly and then running it is very annoying as it does not give the developer any control over what files are being committed.

I've left the index analysis code in the CommitAction since that's necessary for the dialog as the dialog currently a) ignores selection, b) shows "everything", and c) even lets people unselect what to commit. This is unfortunate as I feel dirty about having to pass the collection of un-indexed and un-tracked files into CommitOperation from CommitAction.

Patch-set: 1
Change-id: I719ca671766c5fed471b3b939afdf8cd2291928e
Subject: Refactor CommitAction's code so that a portion of the code can be reused headlessly.
Branch: refs/heads/master
Commit: fe197196cf979de855b1a4b47ce95229c44f97dc